refactored package name to com.slabstech.revive
