+++ bin/main/.gitignore
@@ -1,0 +1,2 @@
+/slabs/
+/com/
