+++ build.gradle
@@ -3,7 +3,7 @@ subprojects {
 
     group = 'com.slabstech.revive'
 
-    version = '0.0.2-SNAPSHOT'
+    version = '2.0.0'
 
     dockerRepo = 'slabstech'
