+++ client/build.gradle
@@ -6,7 +6,7 @@ plugins {
 
 archivesBaseName = 'revive-client'
 
-version = '0.0.2-SNAPSHOT'
+version = '2.0.0'
 
 dockerRepo = 'slabstech'
