+++ server/src/main/java/com/slabstech/revive/persistence/repo/AppUserRepository.java
@@ -1,8 +1,8 @@
-package slabs.tech.avti.persistance.repo;
+package com.slabstech.revive.persistence.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import slabs.tech.avti.persistance.model.AppUser;
+import com.slabstech.revive.persistence.model.AppUser;
 
 public interface AppUserRepository extends JpaRepository<AppUser, Long> {
