+++ server/src/main/java/com/slabstech/revive/persistence/repo/ConfigurationRepository.java
@@ -1,8 +1,8 @@
-package slabs.tech.avti.persistance.repo;
+package com.slabstech.revive.persistence.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import slabs.tech.avti.persistance.model.Configuration;
+import com.slabstech.revive.persistence.model.Configuration;
 
 public interface ConfigurationRepository extends JpaRepository<Configuration, Long> {
