+++ server/src/main/java/com/slabstech/revive/web/AppUserController.java
@@ -1,4 +1,4 @@
-package slabs.tech.avti.web;
+package com.slabstech.revive.web;
 
 import java.util.HashMap;
 import java.util.List;
@@ -18,8 +18,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import slabs.tech.avti.persistance.model.AppUser;
-import slabs.tech.avti.persistance.repo.AppUserRepository;
+import com.slabstech.revive.persistence.model.AppUser;
+import com.slabstech.revive.persistence.repo.AppUserRepository;
+
 import org.springframework.http.HttpStatus;
 
 @RestController
