+++ server/src/main/java/com/slabstech/revive/web/ConfigurationController.java
@@ -1,4 +1,4 @@
-package slabs.tech.avti.web;
+package com.slabstech.revive.web;
 
 import java.util.HashMap;
 import java.util.List;
@@ -18,8 +18,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import slabs.tech.avti.persistance.model.Configuration;
-import slabs.tech.avti.persistance.repo.ConfigurationRepository;
+import com.slabstech.revive.persistence.model.Configuration;
+import com.slabstech.revive.persistence.repo.ConfigurationRepository;
+
 import org.springframework.http.HttpStatus;
 
 @RestController
