+++ server/src/main/java/com/slabstech/revive/web/HomePageController.java
@@ -1,4 +1,4 @@
-package slabs.tech.avti.web;
+package com.slabstech.revive.web;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
