+++ server/src/main/java/com/slabstech/revive/web/exception/DriverIdMismatchException.java
@@ -1,4 +1,4 @@
-package slabs.tech.avti.web.exception;
+package com.slabstech.revive.web.exception;
 
 public class DriverIdMismatchException extends RuntimeException {
