+++ server/src/test/java/com/slabstech/revive/SpringTestConfiguration.java
@@ -1,4 +1,4 @@
-package slabs.tech.avti;
+package com.slabstech.revive;
 
 import org.springframework.context.annotation.Configuration;
